Binary Transformation

Erstellt eine binaere Ergebnismatrix, in der die richtigen Antworten mit einer 1 und die falschen mit einer 0 markiert sind. Falls die richtigen Antworten sowohl Numerics als auch Characters beinhalten, kann der Funktion ebenfalls ein einzeiliger data.frame / tibble gegeben werden. Fehlende Werte werden als falsch beantwortet gewertet.

transform_binary(df, answers)

Arguments

df
Datensatz als data.frame / tibble.

answers
Antwortvektor als Vektor / einzeiliger tibble.

Value

Ergebnismatrix als tibble

Examples

# NOT RUN {
########################
# Identische Datentypen
soscisurvey <- data.frame(
 I1 = c(3, 2, 3, 1),
 I2 = c(2, 2, 2, 1),
 I3 = c(4, 1, 4, 4)
)
antwortvec <- c(3, 2, 4)
transform_binary(soscisurvey, antwortvec)
########################
# Verschiedene Datentypen
soscisurvey <- data.frame(
 I1 = c(3, 2, 3, 1),
 I2 = c("Apfel", "Banane", "Erdbeere", "Apfel"),
 I3 = c(4, 1, NA, 4)
)
antwortvec <- data.frame(3, "Apfel", 4)
transform_binary(soscisurvey, antwortvec)
# }